requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
sentence-transformers==3.1.0
scikit-learn==1.5.2
requests==2.32.3
python-dotenv==1.0.1
numpy==1.26.4
